ユーザ登録処理をcomponent.
